Define ImgraphyType
